refactor: further reduce duplication.
There's still some, but the current level may be acceptable for now. More to come.
