+++ src/main/kotlin/com/autonomousapps/internal/transform/AbstractTransform.kt
@@ -24,6 +24,7 @@ import org.gradle.api.attributes.Category
 @Suppress("UnstableApiUsage")
 internal abstract class AbstractTransform(
   protected val coordinates: Coordinates,
+  protected val declarations: Set<Declaration>,
   protected val configurationNames: ConfigurationNames,
   protected val buildPath: String,
   private val explicitSourceSets: Set<String>,
@@ -71,6 +72,51 @@ internal abstract class AbstractTransform(
         }
       }
     }
+  }
+
+  protected fun addRemainingUsages(usages: Set<Usage>, advice: MutableSet<Advice>) {
+    usages.asSequence()
+      // Don't add unused usages!
+      .filterUsed()
+      // Don't add runtimeOnly or compileOnly (compileOnly, compileOnlyApi, providedCompile) declarations
+      // nb: this probably remains the correct choice, but it can lead to issues when we remove an "unused" dependency
+      // and fail to add a required runtimeOnly dependency that was part of the unused dep's transitive graph. Removing
+      // this line would lead to "super strict" declarations that are, perhaps, more bazel-like (every classpath
+      // consists only of positively-declared dependencies).
+      .filterNot { usage -> usage.bucket == Bucket.COMPILE_ONLY }
+      // Don't add something that is only present on the compileClasspath as that will change the runtimeClasspath,
+      // which we do not want to do.
+      .filterNot { usage ->
+        val identifier = if (coordinates is IncludedBuildCoordinates) {
+          coordinates.resolvedProject.identifier
+        } else {
+          coordinates.identifier
+        }
+
+        val currentClasspaths = dependenciesToClasspaths.get(identifier)
+
+        // if it is a runtime usage and this dep isn't currently in the matching runtime classpath, don't add it there.
+        usage.isRuntimeUsage() && !usage.runtimeMatches(currentClasspaths)
+      }
+      .mapTo(advice) { usage ->
+        val preferredCoordinatesNotation =
+          if (coordinates is IncludedBuildCoordinates && coordinates.resolvedProject.buildPath == buildPath) {
+            coordinates.resolvedProject
+          } else {
+            coordinates
+          }
+        Advice.ofAdd(preferredCoordinatesNotation.withoutDefaultCapability(), mapper.toConfiguration(usage))
+      }
+
+    // Any remaining declarations should be removed
+    declarations.asSequence()
+      // Don't remove runtimeOnly or compileOnly declarations
+      .filterNot { decl ->
+        decl.bucket(configurationNames) == Bucket.COMPILE_ONLY || decl.bucket(configurationNames) == Bucket.RUNTIME_ONLY
+      }
+      .mapTo(advice) { declaration ->
+        Advice.ofRemove(declarationCoordinates(declaration), declaration)
+      }
   }
 
   protected fun Declaration.findSourceKind(hasCustomSourceSets: Boolean): SourceKind? {
